ecs/service/update: Fetch the latest task definition on force
When updating a ECS task definition, We want to deploy a new task with
new revision.
The ECS UpdateService API doesn't update the revision without the
TaskDefinition parameter. To update the revision, we set only a task
family without revision. If a revision is not specified, the latest
ACTIVE revision is used. If the current task uses the latest ACTIVE one,
it hasn't no side-effect.
Since the Force option expects to deploy and converge to the latest
state, when Force option is enabled, we implicitly fetch and set the
task family.